reworked command overwrite implementation
An issue that could occur when using `dev overwrite command` was that if there was a repetitive line of code within the file, the program would just think that that's the start/end point of the script, when it really wasn't. So I fixed that.
Additionally, I fixed an issue where a new line that wasn't supposed to be there appeared when using undoing an overwrite of a command.